requirements.txt

django
djangorestframework
django-notifications-hq
django-filter
djangorestframework-simplejwt
django-extra-fields
django-cors-headers
whitenoise
gunicorn

